Preserve input dtype after LayerNorm (https://github.com/pytorch/pytorch/issues/66707#issuecomment-2028904230)
