+++ sheeprl/utils/model.py
@@ -221,32 +221,3 @@ def cnn_forward(
     flatten_input = input.reshape(-1, *input_dim)
     model_out = model(flatten_input)
     return model_out.reshape(*batch_shapes, *output_dim)
-
-
-class LayerNormChannelLast(nn.LayerNorm):
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-    def forward(self, x: Tensor) -> Tensor:
-        if x.dim() != 4:
-            raise ValueError(f"Input tensor must be 4D (NCHW), received {len(x.shape)}D instead: {x.shape}")
-        x = x.permute(0, 2, 3, 1)
-        x = super().forward(x)
-        x = x.permute(0, 3, 1, 2)
-        return x
-
-
-class LayerNormChannelLastFP32(LayerNormChannelLast):
-    def forward(self, x: Tensor) -> Tensor:
-        input_dtype = x.dtype
-        x = x.to(torch.float32)
-        out = super().forward(x)
-        return out.to(input_dtype)
-
-
-class LayerNormFP32(nn.LayerNorm):
-    def forward(self, x: Tensor) -> Tensor:
-        input_dtype = x.dtype
-        x = x.to(torch.float32)
-        out = super().forward(x)
-        return out.to(input_dtype)
